Fixed backend security and added some frontend logic

+++ todoh/src/main/java/com/hamza/todoh/service/TaskService.java
@@ -1,7 +1,13 @@
 package com.hamza.todoh.service;
 
 import java.util.List;
+import java.util.Optional;
 
+import com.hamza.todoh.auth.service.CustomUserDetailsService;
+import com.hamza.todoh.model.User;
+import com.hamza.todoh.repository.UserRepository;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import com.hamza.todoh.dto.TaskRequestDto;
@@ -22,11 +28,25 @@ public class TaskService {
 
     private final TaskMapperDto taskMapperDto;
 
+    private final UserRepository userRepository;
+
+
     
 
     public TaskResponseDto delete(Integer id) {
         
         var task = taskRepository.findById(id).orElseThrow(() -> new RuntimeException("Task not found"));
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        String username = authentication.getName();
+
+
+        if(!task.getUser().getUsername().equals(username)) {
+            throw  new RuntimeException("Shit aint yours ");
+        }
+
+
         taskRepository.delete(task);
         return taskMapperDto.mapToTaskResponseDto(task);
     }
@@ -35,17 +55,54 @@ public class TaskService {
 
     public TaskResponseDto getTask(Integer id) {
         var task = taskRepository.findById(id).orElseThrow(() -> new RuntimeException("Task not found"));
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        String username = authentication.getName();
+
+
+        if(!task.getUser().getUsername().equals(username)) {
+            throw  new RuntimeException("Shit aint yours ");
+        }
+
+
+
         return taskMapperDto.mapToTaskResponseDto(task);
+
+
+
     }
 
     public TaskResponseDto addTask(TaskRequestDto dto) {
-        var task = taskRepository.save(taskMapperDto.mapToTask(dto));
-        return taskMapperDto.mapToTaskResponseDto(task);
+
+
+        var task = taskMapperDto.mapToTask(dto);
+
+
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        var user = userRepository.findByUsername(authentication.getName());
+        task.setUser(user.get());
+
+        var savedTask= taskRepository.save(task);
+        return taskMapperDto.mapToTaskResponseDto(savedTask); //TODO cambiare la funzione
     }
 
 
     public List<TaskResponseDto> getTasks() {
-        return taskRepository.findAll().stream().map(taskMapperDto::mapToTaskResponseDto).toList();
+
+
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        String username = authentication.getName();
+
+        Optional<User> user = Optional.ofNullable(userRepository.findByUsername(username).orElseThrow(() -> new RuntimeException("User not found ")));
+
+
+
+        return taskRepository.findByUser(user.get()).stream().map(taskMapperDto::mapToTaskResponseDto).toList();
     }
